(PC-35563) refactor: extract deposit api from finance api
