+++ api/src/pcapi/core/users/api.py
@@ -35,8 +35,8 @@ from pcapi.core.chronicles import constants as chronicles_constants
 from pcapi.core.chronicles import models as chronicles_models
 from pcapi.core.external.attributes import api as external_attributes_api
 from pcapi.core.external.sendinblue import update_contact_attributes
+from pcapi.core.finance import deposit_api
 from pcapi.core.finance import models as finance_models
-import pcapi.core.finance.api as finance_api
 from pcapi.core.fraud import models as fraud_models
 import pcapi.core.fraud.api as fraud_api
 import pcapi.core.fraud.common.models as common_fraud_models
@@ -757,7 +757,7 @@ def _update_underage_beneficiary_deposit_expiration_date(user: models.User) -> N
         raise ValueError("Trying to update underage beneficiary deposit expiration date but user has no deposit")
 
     current_deposit_expiration_datetime = user.deposit.expirationDate
-    new_deposit_expiration_datetime = finance_api.compute_underage_deposit_expiration_datetime(user.birth_date)
+    new_deposit_expiration_datetime = deposit_api.compute_underage_deposit_expiration_datetime(user.birth_date)
 
     if current_deposit_expiration_datetime == new_deposit_expiration_datetime:
         return
@@ -919,7 +919,7 @@ def create_pro_user(pro_user: users_serialization.ProUserCreationBodyV2Model) ->
         eighteen_years_ago = datetime.datetime.utcnow() - datetime.timedelta(days=366 * 18)
         new_pro_user.dateOfBirth = eighteen_years_ago
         new_pro_user.validatedBirthDate = new_pro_user.dateOfBirth.date()
-        deposit = finance_api.create_deposit(new_pro_user, "integration_signup", models.EligibilityType.AGE18)
+        deposit = deposit_api.create_deposit(new_pro_user, "integration_signup", models.EligibilityType.AGE18)
         new_pro_user.deposits = [deposit]
 
     db.session.add(new_pro_user)
