+++ api/tests/sandboxes/scripts/create_industrial_app_users_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pcapi.core.finance.api import recredit_users
+from pcapi.core.finance.deposit_api import recredit_users
 import pcapi.core.finance.models as finance_models
 from pcapi.core.users.models import User
 from pcapi.core.users.models import UserRole
